portions normalizer
