+++ ydb/core/tx/columnshard/data_sharing/protos/data.proto
@@ -26,12 +26,13 @@ message TIndexChunk {
 message TPortionInfo {
     optional uint64 PathId = 1;
     optional uint64 PortionId = 2;
-    optional NKikimrColumnShardProto.TSnapshot MinSnapshot = 3;
+    optional NKikimrColumnShardProto.TSnapshot MinSnapshotDeprecated = 3;
     optional NKikimrColumnShardProto.TSnapshot RemoveSnapshot = 4;
     optional NKikimrTxColumnShard.TIndexPortionMeta Meta = 5;
     repeated TColumnRecord Records = 6;
     repeated TIndexChunk Indexes = 7;
     repeated NKikimrColumnShardProto.TUnifiedBlobId BlobIds = 8;
+    optional uint64 SchemaVersion = 9;
 }
 
 message TPathIdData {
