+++ ydb/core/tx/columnshard/normalizer/portion/normalizer.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "normalizer.h"
+
 #include <ydb/core/tx/columnshard/normalizer/abstract/abstract.h>
 #include <ydb/core/tx/columnshard/engines/scheme/abstract_scheme.h>
 #include <ydb/core/tx/columnshard/blobs_action/counters/storage.h>
@@ -75,4 +77,27 @@ public:
                     std::make_shared<TReadPortionsTask<TConveyorTask>>(nCtx, actions, std::move(Package), Schemas), 1, memSize, "CS::NORMALIZER", controller.GetTaskSubscription()));
     }
 };
+
+class TPortionsNormalizerBase : public INormalizerComponent {
+public:
+    TPortionsNormalizerBase(TTabletStorageInfo* info)
+        : DsGroupSelector(info)
+    {}
+
+    virtual TConclusion<std::vector<INormalizerTask::TPtr>> Init(const TNormalizationController& controller, NTabletFlatExecutor::TTransactionContext& txc) override final;
+
+protected:
+    virtual INormalizerTask::TPtr BuildTask(std::vector<std::shared_ptr<TPortionInfo>>&& portions, std::shared_ptr<THashMap<ui64, ISnapshotSchema::TPtr>> schemas) const = 0;
+    virtual TConclusion<bool> DoInit(const TNormalizationController& controller, NTabletFlatExecutor::TTransactionContext& txc)  = 0;
+
+    virtual bool CheckPortion(const TPortionInfo& /*portionInfo*/) const = 0;
+
+    virtual std::set<ui32> GetColumnsFilter(const ISnapshotSchema::TPtr& schema) const {
+        return schema->GetPkColumnsIds();
+    }
+
+private:
+    NColumnShard::TBlobGroupSelector DsGroupSelector;
+};
+
 }
